+++ helpers/index.js
@@ -1,1 +1,7 @@
 const HttpError = require("./HttpError");
+const ctrlWrapper = require("./ctrlWrapper");
+
+module.exports = {
+  HttpError,
+  ctrlWrapper,
+};
