+++ middlewares/index.js
@@ -1,0 +1,2 @@
+const validateBody = require("./validateBody");
+module.exports = validateBody;
