+++ models/contacts.json
@@ -59,7 +59,6 @@
     "email": "dui.Fusce.diam@Donec.com",
     "phone": "(233) 738-2360"
   },
-
   {
     "id": "WV9PjPDS8M2PNO8W8X_eH",
     "name": "Bob Dylon",
